requirements.txt
streamlit==1.28.1
openai-whisper
pymongo==4.6.0
boto3==1.34.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
streamlit-player==0.1.5
moviepy==1.0.3
pydub==0.25.1
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.25.2 